MimeType.schema set to value in supplied raml
This is useful when we want to use a json schema that contains
references other schemas in different files. When we just have the
schema as a raw string, including something like "./schema.js" doesn't
work since there is not base path to look in.

The validation rule is update to reflect this change, and will either
get the InputStream from the external file or use the string itself if
it is json.

I can see that the url is located in IncludeScalarNode, so it exists
while parsing, but it is just a matter of figuring out how to get into
the MimeType object.